validate endpoints and handle mismatched type exceptions

+++ src/main/java/com/antarctica/explorer/api/repository/CruiseLineRepository.java
@@ -4,6 +4,6 @@ import com.antarctica.explorer.api.model.CruiseLine;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CruiseLineRepository extends JpaRepository<CruiseLine, Long> {
+public interface CruiseLineRepository extends JpaRepository<CruiseLine, Integer> {
   Optional<CruiseLine> findByName(String name);
 }
